requirements.txt
tensorflow-cpu==2.8.0
matplotlib==3.2.2
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
